Fix for task 1-- truncating file instead of appending

+++ 0x15-file_io/1-create_file.c
@@ -9,7 +9,7 @@
 */
 int create_file(const char *filename, char *text_content)
 {
-	int fd, len = 0, writelen;
+	long int fd, len = 0, writelen;
 
 	if (filename == NULL)
 		return (-1);
